requirements.txt
imbalanced_learn==0.12.0
joblib==1.2.0
openai==1.3.4
langchain==0.1.6
matplotlib==3.7.2
nltk==3.8.1
numpy==1.24.3
plotly==5.18.0
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
streamlit==1.31.0
streamlit_lottie==0.0.5
wordcloud==1.9.3
xgboost==2.0.3
statsmodels==0.14.0
tiktoken
langchain-community
pdfplumber
python-dotenv
replicate
streamlit_chat
pandas==1.5.3
seaborn==0.13.2
statsmodels==0.14.0
pandasai
sqlalchemy
io